requirements.txt
mysql-connector-python
folium